refactor(#2826): move View into Explorer package
