+++ lua/nvim-tree/explorer/view.lua
